refactor(website): align product-stepper structure with product-feature
Give each scene visual its own self-contained folder (DataModelVisual,
WorkflowVisual, LayoutVisual), each with an index.ts barrel and
components/ data/ types/ (+ utils/ for workflow) — mirroring the
product-feature convention. Split every domain type into its own file
under types/ (one type per file); each data file now has a single value
export. The shared shell (ProductStepperShell/Content/Visual/VisualFrame)
and the section steps data stay at the section root.

Structure only — all moves are git renames, no behavior change.
