+++ packages/twenty-website/src/sections/product-stepper/DataModelVisual/index.ts
@@ -1,0 +1,1 @@
+export { DataModelVisual } from './DataModelVisual';
