+++ packages/twenty-website/src/sections/product-stepper/DataModelVisual/types/data-model-field-icon.ts
@@ -1,0 +1,6 @@
+export type DataModelFieldIcon =
+  | 'apps'
+  | 'building'
+  | 'tag'
+  | 'target'
+  | 'user';
