+++ packages/twenty-website/src/sections/product-stepper/DataModelVisual/types/data-model-header-icon.ts
@@ -1,0 +1,7 @@
+export type DataModelHeaderIcon =
+  | 'briefcaseSmall'
+  | 'buildingSmall'
+  | 'targetSmall'
+  | 'userScreenSmall'
+  | 'userSmall'
+  | 'usersSmall';
