+++ packages/twenty-website/src/sections/product-stepper/DataModelVisual/types/entity-connection.ts
@@ -1,0 +1,4 @@
+export type EntityConnection = {
+  from: string;
+  to: string;
+};
