+++ packages/twenty-website/src/sections/product-stepper/DataModelVisual/types/entity-tone.ts
@@ -1,0 +1,3 @@
+import { type PRODUCT_STEPPER_SCENE } from '@/tokens/feature-scenes/product-stepper-scene';
+
+export type EntityTone = keyof typeof PRODUCT_STEPPER_SCENE.entityTones;
