+++ packages/twenty-website/src/sections/product-stepper/LayoutVisual/components/layout-chrome.ts
