+++ packages/twenty-website/src/sections/product-stepper/LayoutVisual/data/layout-data.ts
@@ -1,36 +1,8 @@
 import { type MessageDescriptor } from '@lingui/core';
 import { msg } from '@lingui/core/macro';
 
-import { type PRODUCT_STEPPER_SCENE } from '@/tokens/feature-scenes/product-stepper-scene';
-
-import {
-  type LayoutFieldIconType,
-  type LayoutNavIconType,
-} from '../components/LayoutIcons';
-
-type NavColor = keyof typeof PRODUCT_STEPPER_SCENE.navTiles;
-
-export type LayoutFieldDefinition = {
-  icon: LayoutFieldIconType;
-  id: string;
-  label: MessageDescriptor;
-  section: string;
-  type: MessageDescriptor;
-  visible: boolean;
-};
-
-export type LayoutNavItemDefinition = {
-  children?: {
-    color: NavColor;
-    icon: LayoutNavIconType;
-    label: MessageDescriptor;
-  }[];
-  color: NavColor;
-  icon: LayoutNavIconType;
-  isActive: boolean;
-  isFolder?: boolean;
-  label: MessageDescriptor;
-};
+import { type LayoutFieldDefinition } from '../types/layout-field-definition';
+import { type LayoutNavItemDefinition } from '../types/layout-nav-item-definition';
 
 export const LAYOUT_EDITOR_CONTENT: {
   fields: LayoutFieldDefinition[];
