+++ packages/twenty-website/src/sections/product-stepper/LayoutVisual/index.ts
@@ -1,0 +1,1 @@
+export { LayoutVisual } from './LayoutVisual';
