+++ packages/twenty-website/src/sections/product-stepper/LayoutVisual/types/layout-field-definition.ts
@@ -1,0 +1,12 @@
+import { type MessageDescriptor } from '@lingui/core';
+
+import { type LayoutFieldIconType } from './layout-field-icon-type';
+
+export type LayoutFieldDefinition = {
+  icon: LayoutFieldIconType;
+  id: string;
+  label: MessageDescriptor;
+  section: string;
+  type: MessageDescriptor;
+  visible: boolean;
+};
