+++ packages/twenty-website/src/sections/product-stepper/LayoutVisual/types/layout-field-icon-type.ts
@@ -1,0 +1,8 @@
+export type LayoutFieldIconType =
+  | 'calendar'
+  | 'link'
+  | 'map'
+  | 'money'
+  | 'target'
+  | 'user'
+  | 'users';
