+++ packages/twenty-website/src/sections/product-stepper/LayoutVisual/types/layout-nav-icon-type.ts
@@ -1,0 +1,10 @@
+export type LayoutNavIconType =
+  | 'automation'
+  | 'building'
+  | 'checkbox'
+  | 'dashboard'
+  | 'history'
+  | 'notes'
+  | 'target'
+  | 'user'
+  | 'versions';
