+++ packages/twenty-website/src/sections/product-stepper/WorkflowVisual/index.ts
@@ -1,0 +1,1 @@
+export { WorkflowVisual } from './WorkflowVisual';
