+++ packages/twenty-website/src/sections/product-stepper/WorkflowVisual/types/workflow-edge-definition.ts
@@ -1,0 +1,4 @@
+export type WorkflowEdgeDefinition = {
+  from: string;
+  to: string;
+};
