+++ packages/twenty-website/src/sections/product-stepper/WorkflowVisual/types/workflow-icon-name.ts
@@ -1,0 +1,8 @@
+export type WorkflowIconName =
+  | 'brain'
+  | 'filter'
+  | 'playlistAdd'
+  | 'plus'
+  | 'reload'
+  | 'search'
+  | 'send';
