+++ packages/twenty-website/src/sections/product-stepper/WorkflowVisual/utils/use-workflow-animation.ts
