+++ packages/twenty-website/src/sections/product-stepper/types/product-stepper-visual-key.ts
@@ -1,0 +1,1 @@
+export type ProductStepperVisualKey = 'dataModel' | 'layout' | 'workflow';
